fix: pointMapper 제거
